# config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Ecommerce",
    "version": "1.0.0",
    "description": "Documentación de la API para la tienda Ecommerce"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Servidor de Desarrollo"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID único de la categoría",
            "example": 1
          },
          "nombre": {
            "type": "string",
            "description": "Nombre de la categoría",
            "example": "Electrónica"
          },
          "descripcion": {
            "type": "string",
            "description": "Descripción de la categoría",
            "example": "Productos electrónicos y gadgets."
          },
          "imagen": {
            "type": "string",
            "description": "Nombre del archivo de imagen de la categoría",
            "example": "electronics.jpg"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/categories": {
      "get": {
        "summary": "Obtener todas las categorías",
        "tags": [
          "Categories"
        ],
        "responses": {
          "200": {
            "description": "Lista de categorías",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error al obtener las categorías"
          }
        }
      },
      "post": {
        "summary": "Crear una nueva categoría (Solo Administradores)",
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "nombre"
                ],
                "properties": {
                  "nombre": {
                    "type": "string",
                    "description": "Nombre de la categoría"
                  },
                  "descripcion": {
                    "type": "string",
                    "description": "Descripción de la categoría"
                  },
                  "imagen": {
                    "type": "string",
                    "format": "binary",
                    "description": "Imagen de la categoría"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Categoría creada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "400": {
            "description": "Error de validación"
          },
          "403": {
            "description": "Acceso denegado"
          },
          "500": {
            "description": "Error al crear la categoría"
          }
        }
      }
    },
    "/categories/{id}": {
      "put": {
        "summary": "Actualizar una categoría existente (Solo Administradores)",
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID de la categoría a actualizar"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string",
                    "description": "Nombre de la categoría"
                  },
                  "descripcion": {
                    "type": "string",
                    "description": "Descripción de la categoría"
                  },
                  "imagen": {
                    "type": "string",
                    "format": "binary",
                    "description": "Nueva imagen de la categoría"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Categoría actualizada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "400": {
            "description": "Error de validación"
          },
          "403": {
            "description": "Acceso denegado"
          },
          "404": {
            "description": "Categoría no encontrada"
          },
          "500": {
            "description": "Error al actualizar la categoría"
          }
        }
      },
      "delete": {
        "summary": "Eliminar una categoría existente (Solo Administradores)",
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID de la categoría a eliminar"
          }
        ],
        "responses": {
          "200": {
            "description": "Categoría eliminada exitosamente"
          },
          "403": {
            "description": "Acceso denegado"
          },
          "404": {
            "description": "Categoría no encontrada"
          },
          "500": {
            "description": "Error al eliminar la categoría"
          }
        }
      }
    },
...
  },
...
}